requirements.txt
Flask==2.3.2
Flask-WTF==1.1.1
Werkzeug==2.3.6
Jinja2==3.1.2
nltk==3.8.1
PyPDF2==3.0.1
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
